requirements.txt
beautifulsoup4==4.10.0
requests==2.27.1
SQLAlchemy==1.4.31
XlsxWriter==3.0.2
fpdf2==2.5.0